Small app folder tweaks - Restructured app/pages folder - Restructured app/features folder - Added numerous index.ts files to folders under app/
