+++ projects/app/src/app/features/footers/index.ts
@@ -1,1 +1,3 @@
+export * from './components';
+
 export * from './footers.module';
