+++ projects/app/src/app/features/headers/components/header-default/header-default.component.scss
