+++ projects/app/src/app/features/index.ts
@@ -1,0 +1,3 @@
+export * from './footers';
+export * from './headers';
+export * from './layouts';
