+++ projects/app/src/app/features/layouts/layouts.module.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
-import { LayoutDefaultComponent } from './layout-default/layout-default.component';
+import { LayoutDefaultComponent } from './components';
 
 const exportedDeclarations: any[] = [
   LayoutDefaultComponent
