+++ projects/app/src/app/pages/dashboard/components/dashboard/dashboard.component.scss
