+++ projects/app/src/app/pages/dashboard/components/index.ts
@@ -1,0 +1,1 @@
+export * from './dashboard/dashboard.component';
