+++ projects/app/src/app/pages/dashboard/index.ts
@@ -1,3 +1,4 @@
+export * from './components';
+
 export * from './dashboard-routing.module';
-export * from './dashboard.component';
 export * from './dashboard.module';
