+++ projects/app/src/environments/index.ts
@@ -1,0 +1,2 @@
+export * from './environment.model';
+export * from './environment';
